change to long int

+++ _printf.c
@@ -12,7 +12,7 @@
 int _printf(const char *format, ...)
 {
 	int h = 0;
-	int j = 0;
+	long int j = 0;
 	int i = 0;
 	int n = 1;
 	int *cont = &n;
